cucumber-expressions: limit explosion of generated expressions to 256
Using only small input values it is possible make
`CucumberExpressionGenerator` generate 1.5GB worth of generated
expressions.

For example: the step `a simple step` -a total of 11 letters- and 4
parameters with a the regular expression of [a-z] will generate a
4^11=419430 possible cucumber expressions. As each GeneratedExpression
takes up approximately 11*32 bytes this totals just shy of 1.5GB.

While this can be avoided by not using parameter types with overly
generic expressions for snippet generation, snippet generation is
enabled by default and once we run out memory it is impossible to
inform or warn anyone about this problem.

So 256 cucumber expressions should be enough for anyone.

Fixes: #574