cucumber-expressions: Prefer expression with the longest non-empty match
Given `a simple step` and a parameter type that matches zero or more
characters `[a-z]*` and a parameter type that matches exactly one
character `[a-z]` two expressions will be generated. Neither very useful.

```
{exactly-one} {exactly-one}{zero-or-more} {exactly-one}{zero-or-more}
{zero-or-more} {exactly-one}{zero-or-more} {exactly-one}{zero-or-more}
```

This result is generated because after advancing `ParameterTypeMatcher`
to just after `a` the zero-or-more expression will match the empty
string between `a` and the following space. The `exactly-one-expression`
will proceed and match the first character after that space.

By advancing `ParameterTypeMatcher` until a non empty match is
found we can avoid expressions that match the empty string from getting
stuck.

This will generate:

```
{exactly-one} {zero-or-more} {zero-or-more}
{zero-or-more} {zero-or-more} {zero-or-more}
```